Ensure the original user in the IMBO url is maintained when resizing
When resizing, a new URL is created with the username from the configuration. This causes problems for VGNett, as IMBO-images may have multiple sources. Hardcoding the user results in imag

This change ensures the username from the original URL is kept when we generate the new URL. The originally configured user is reset once the URL has been generated.